Remove / disable old sharing code
