+++ appinfo/app.php
@@ -31,5 +31,4 @@ require_once __DIR__ . '/autoload.php';
 $app = new Application(); // \AppInfo\Application();
 $app->registerNavigationEntry();
 
-\OC\Share\Share::registerBackend ('nextnote', '\OCA\NextNote\ShareBackend\NextNoteShareBackend');
-\OCP\App::registerAdmin('nextnote', 'admin');
+
