+++ controller/notebookapicontroller.php
@@ -24,36 +24,29 @@
 namespace OCA\NextNote\Controller;
 
 use OCA\NextNote\Db\Notebook;
-use OCA\NextNote\Fixtures\ShareFix;
 use OCA\NextNote\Service\NotebookService;
-use OCA\NextNote\Service\NoteService;
-use OCA\NextNote\ShareBackend\NextNoteShareBackend;
 use OCA\NextNote\Utility\NotFoundJSONResponse;
 use OCA\NextNote\Utility\UnauthorizedJSONResponse;
 use OCA\NextNote\Utility\Utils;
 use \OCP\AppFramework\ApiController;
 use OCP\AppFramework\Http\JSONResponse;
-use OCP\Constants;
 use OCP\IConfig;
 use OCP\ILogger;
 use \OCP\IRequest;
 use OCP\IUserManager;
-use OCP\Share;
 
 
 class NotebookApiController extends ApiController {
 
 	private $config;
 	private $notebookService;
-	private $shareBackend;
 	private $userManager;
 
 	public function __construct($appName, IRequest $request,
-								ILogger $logger, IConfig $config, NotebookService $notebookService, NextNoteShareBackend $shareBackend, IUserManager $userManager) {
+								ILogger $logger, IConfig $config, NotebookService $notebookService, IUserManager $userManager) {
 		parent::__construct($appName, $request);
 		$this->config = $config;
 		$this->notebookService = $notebookService;
-		$this->shareBackend = $shareBackend;
 		$this->userManager = $userManager;
 	}
