+++ templates/main.php
@@ -40,7 +40,7 @@ script('nextnote', 'vendor/angular-translate/angular-translate-loader-url.min');
 script('nextnote', 'app/app');
 script('nextnote', 'app/routes');
 script('nextnote', 'templates');
-script('nextnote', 'share');
+script('nextnote', 'animations');
 script('nextnote', 'app/controllers/MainCtrl');
 script('nextnote', 'app/controllers/NoteListCtrl');
 script('nextnote', 'app/controllers/NoteViewCtrl');
