+++ templates/views/list.html
@@ -85,29 +85,29 @@
 							<span class="pointer buttons edit"
 								  tooltip="{{ 'edit.note' | translate }}"></span>
 							</div>
-							<div id="note-{{ note.id }}-share"
-								 class="share-note share pointer"
-								 ng-click="shareNote(note)">
-								<a class="share" data-item-type="nextnote"
-								   data-item="{{ note.id }}"
-								   data-possible-permissions="{{ note.permissions }}"
-								   data-path="{{ note.id }}">
+							<!--<div id="note-{{ note.id }}-share"-->
+								 <!--class="share-note share pointer"-->
+								 <!--ng-click="shareNote(note)">-->
+								<!--<a class="share" data-item-type="nextnote"-->
+								   <!--data-item="{{ note.id }}"-->
+								   <!--data-possible-permissions="{{ note.permissions }}"-->
+								   <!--data-path="{{ note.id }}">-->
 
-									<div class="buttons share shared share-note pointer"
-										 ng-if="(note.shared_with | objectKeysLength) > 0 && note.owner.uid === OC.currentUser">
-										Shared
-									</div>
-									<div class="buttons share shared share-note pointer"
-										 ng-if="(note.shared_with | objectKeysLength) > 0 && note.owner.uid !== OC.currentUser"
-										 tooltip="{{ 'sharedwith' | translate:{name: note.owner.display_name } }}">
-										{{ note.owner.display_name }}
-									</div>
-									<div id="note-{{ note.id }}"
-										 class="buttons share share-note pointer"
-										 tooltip="{{ 'share.note' | translate }}"
-										 ng-if="note.shared_with.length === 0"></div>
-								</a>
-							</div>
+									<!--<div class="buttons share shared share-note pointer"-->
+										 <!--ng-if="(note.shared_with | objectKeysLength) > 0 && note.owner.uid === OC.currentUser">-->
+										<!--Shared-->
+									<!--</div>-->
+									<!--<div class="buttons share shared share-note pointer"-->
+										 <!--ng-if="(note.shared_with | objectKeysLength) > 0 && note.owner.uid !== OC.currentUser"-->
+										 <!--tooltip="{{ 'sharedwith' | translate:{name: note.owner.display_name } }}">-->
+										<!--{{ note.owner.display_name }}-->
+									<!--</div>-->
+									<!--<div id="note-{{ note.id }}"-->
+										 <!--class="buttons share share-note pointer"-->
+										 <!--tooltip="{{ 'share.note' | translate }}"-->
+										 <!--ng-if="note.shared_with.length === 0"></div>-->
+								<!--</a>-->
+							<!--</div>-->
 						</div>
 					</td>
 					<td class="actions" ng-if="list_filter.deleted === 1">
